requirements.txt
streamlit==1.25.0
torch==2.3.1+cu121
diffusers==0.22.1
Pillow==10.0.1
matplotlib==3.7.2
